+++ docs/Changelog.md
@@ -11,6 +11,7 @@
 * [#3029](https://github.com/TouK/nussknacker/pull/3029) Added `kafka.schemaRegistryCacheConfig` (was hardcoded before)
 * [#3047](https://github.com/TouK/nussknacker/pull/3047) Remove deprecated Admin panel tabs that are replaced with Components tab: 
   Search Components and Unused Components (together with API endpoints: /processesComponents and /unusedComponents)
+* [#3049](https://github.com/TouK/nussknacker/pull/3049) Added `collector` component to lite base components
 
 1.3.1 (Not released yet)
 ------------------------